Move planning code into new file
