+++ src/main/java/com/android/volley/toolbox/BuildableRequest.java
@@ -1,59 +1,18 @@
 package com.android.volley.toolbox;
 
 import com.android.volley.AuthFailureError;
-import com.android.volley.DefaultRetryPolicy;
 import com.android.volley.NetworkResponse;
 import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 
-import org.json.JSONObject;
-
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 /**
  * TODO
  */
 public class BuildableRequest<T> extends Request<T> {
-
-    /**
-     * TODO remove this (this is just for planning)
-     */
-    public static void example() {
-        Response.Listener<Object> listener = null;
-        Response.ErrorListener errorListener = null;
-        //noinspection ConstantConditions
-        RequestBuilder.create()
-                .template(Templators.forJSON())
-                .url("url") // TODO null check
-                .appendUrl("url") // TODO Throw exception if no url yet
-                .method(Method.GET) // TODO null check, or default to get?
-                .header("Key", "Val")
-                .header("Key2", "Val2")
-                .headers(new HashMap<String, String>())
-                .range("name", 0, 99)
-                .rangeForPage("name", 0, 100)
-                .param("Key", "Val")
-                .params(new HashMap<String, String>())
-                .paramsEncoding("")
-                .body(Bodies.forJSONObject(new JSONObject())) // todo also Bodies.forParams(map)
-                .bodyContentType(BodyContents.forJSON()) // todo include this in the body,
-                .parseResponse(ResponseParsers.forJSONObject()) // todo force generic, don't allow re changing
-                .onSuccess(listener) // TODO parsing cannot be done after adding listeners
-                .onError(errorListener) // TODO null check, don't accept a list
-                .priority(Priority.NORMAL)
-                .tag("tag")
-                .marker("debugMarker")
-                .retryPolicy(new DefaultRetryPolicy())
-                .retryOnServerErrors(true)
-                .shouldCache(true)
-                .build() // TODO don't build twice
-                .addTo(Volley.newRequestQueue(null));
-    }
 
     private final ResponseParser<T> parser;
     private final String bodyContentType;
